Add BPKNavigationView+Style.swift
